+++ activerecord/test/cases/query_logs_test.rb
@@ -204,7 +204,7 @@ class QueryLogsTest < ActiveRecord::TestCase
     ActiveRecord::QueryLogs.tags_formatter = :legacy
     ActiveRecord::QueryLogs.tags = [ :application ]
 
-    with_query_log_tags_format("sqlcommenter") do |connection|
+    with_db_config(query_log_tags_format: "sqlcommenter") do |connection|
       assert_equal "select 1 /*application='active_record'*/",
         ActiveRecord::QueryLogs.call("select 1", connection)
     end
@@ -214,7 +214,7 @@ class QueryLogsTest < ActiveRecord::TestCase
     ActiveRecord::QueryLogs.tags_formatter = :sqlcommenter
     ActiveRecord::QueryLogs.tags = [ :application ]
 
-    with_query_log_tags_format(nil) do |connection|
+    with_db_config do |connection|
       assert_equal "select 1 /*application='active_record'*/",
         ActiveRecord::QueryLogs.call("select 1", connection)
     end
@@ -225,14 +225,44 @@ class QueryLogsTest < ActiveRecord::TestCase
     ActiveRecord::QueryLogs.tags_formatter = :legacy
     ActiveRecord::QueryLogs.tags = [ :application ]
 
-    with_query_log_tags_format(nil) do |legacy_connection|
-      with_query_log_tags_format("sqlcommenter") do |sqlcommenter_connection|
+    with_db_config do |legacy_connection|
+      with_db_config(query_log_tags_format: "sqlcommenter") do |sqlcommenter_connection|
         # Different formats must not serve each other's cached comment.
         assert_equal "select 1 /*application:active_record*/",
           ActiveRecord::QueryLogs.call("select 1", legacy_connection)
         assert_equal "select 1 /*application='active_record'*/",
           ActiveRecord::QueryLogs.call("select 1", sqlcommenter_connection)
       end
+    end
+  end
+
+  def test_per_pool_prepend_comment_overrides_global
+    ActiveRecord::QueryLogs.prepend_comment = false
+    ActiveRecord::QueryLogs.tags = [ :application ]
+
+    with_db_config(query_log_tags_prepend_comment: true) do |connection|
+      assert_equal "/*application:active_record*/ select 1",
+        ActiveRecord::QueryLogs.call("select 1", connection)
+    end
+  end
+
+  def test_per_pool_prepend_comment_falls_back_to_global_when_not_set
+    ActiveRecord::QueryLogs.prepend_comment = true
+    ActiveRecord::QueryLogs.tags = [ :application ]
+
+    with_db_config do |connection|
+      assert_equal "/*application:active_record*/ select 1",
+        ActiveRecord::QueryLogs.call("select 1", connection)
+    end
+  end
+
+  def test_per_pool_prepend_comment_can_disable_global
+    ActiveRecord::QueryLogs.prepend_comment = true
+    ActiveRecord::QueryLogs.tags = [ :application ]
+
+    with_db_config(query_log_tags_prepend_comment: false) do |connection|
+      assert_equal "select 1 /*application:active_record*/",
+        ActiveRecord::QueryLogs.call("select 1", connection)
     end
   end
 
@@ -326,10 +356,9 @@ class QueryLogsTest < ActiveRecord::TestCase
   end
 
   private
-    def with_query_log_tags_format(format, &block)
+    def with_db_config(options = {}, &block)
       base_config = ActiveRecord::Base.connection_pool.db_config
-      configuration_hash = base_config.configuration_hash
-      configuration_hash = configuration_hash.merge(query_log_tags_format: format) if format
+      configuration_hash = base_config.configuration_hash.merge(options)
       db_config = ActiveRecord::DatabaseConfigurations::HashConfig.new(base_config.env_name, base_config.name, configuration_hash)
 
       pool = ActiveRecord::ConnectionAdapters::PoolConfig.new(ActiveRecord::Base, db_config, :writing, :default).pool